more comments about u32
