+++ codec/src/types/mod.rs
@@ -1,3 +1,6 @@
+//! Codec implementations for common types
+
+pub mod bytes;
 pub mod net;
 pub mod primitives;
 pub mod tuple;
