+++ codec/src/types/net.rs
@@ -1,3 +1,5 @@
+//! Codec implementations for network-related types
+
 use crate::{EncodeSize, Error, FixedSize, Read, ReadExt, Write};
 use bytes::{Buf, BufMut};
 use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
